Runtime extractions, bug fix & window store
Runtime:
- Add store for the global window object.
- Method access store from the metalib generated file, which will determine the store for the environment(nodejs/browser) and return the store instance
- Extract descriptions/classes from the main Reflect file

Transformer:
- Fix line I missed, which makes the transformer get the correct stub file for the writer
- Remove MetaStoreImpl.ts(was not used/is' not needed)
- Fix issue where meta lib is generated in root dir, rather than src dir and transpiled into dist dir (probably need a better solution for this, incase no root/dist dir is defined)
