+++ runtime/src/const.ts
@@ -1,0 +1,18 @@
+/**
+ * To identify functions by package
+ */
+export const TYPE_ID_PROPERTY_NAME = "__tst_reflect__";
+/**
+ * Name of JSDoc comment marking method as it use type of generic parameter.
+ */
+export const REFLECT_GENERIC_DECORATOR = "reflectGeneric";
+/**
+ * Name of JSDoc comment marking function as decorator.
+ * @type {string}
+ */
+export const REFLECT_DECORATOR_DECORATOR = "reflectDecorator";
+/**
+ * Name of the getType() function
+ */
+export const GET_TYPE_FNC_NAME = "getType";
+export const REFLECT_STORE_SYMBOL = Symbol('tst_reflect_store');
