+++ runtime/src/descriptions/index.ts
@@ -1,0 +1,27 @@
+import { ConditionalType } from "./conditional-type";
+import { Decorator }       from "./decorator";
+import {
+	Constructor,
+	ConstructorActivator
+}                          from "./method";
+import { MethodParameter } from "./parameter";
+import { Property }        from "./property";
+
+export {
+	MethodBase,
+	Method,
+	MethodActivator
+}                               from './method';
+export { Property }          from './property';
+export { Decorator }            from './decorator';
+export { MethodParameter }      from './parameter';
+export {
+	ConditionalType
+}                               from './conditional-type';
+export {
+	ConstructorImport,
+	ConstructorImportActivator
+}                               from './constructor-import';
+export { IndexedAccessType }    from './indexed-access-type';
+export { ConstructorActivator } from "./method";
+export { Constructor }          from "./method";
