+++ runtime/src/reflect.ts
@@ -1,764 +1,26 @@
 /* eslint-disable */
 
 
-import { MetaStoreImpl } from "./meta-stores/MetaStoreImpl";
-import { InlineStore }   from "./meta-stores/InlineStore";
-/**
- * Kind of type
- */
-export enum TypeKind
-{
-	/**
-	 * Interface
-	 */
-	Interface,
+import {
+	ConditionalType,
+	Constructor,
+	ConstructorImport,
+	ConstructorImportActivator,
+	Decorator,
+	IndexedAccessType,
+	Method,
+	MethodParameter,
+	Property,
+}                         from "./descriptions";
+import {
+	Mapper,
+	resolveLazyType
+}                         from "./mapper";
+import { TypeProperties } from "./descriptions/type-properties";
+import { TypeKind }       from "./flags";
+import { InlineStore }    from "./meta-stores/InlineStore";
+import { MetaStoreImpl }  from "./meta-stores/MetaStoreImpl";
 
-	/**
-	 * Class
-	 */
-	Class,
-
-	/**
-	 * Native JavaScript/TypeScript type
-	 */
-	Native,
-
-	/**
-	 * Container for other types in case of types union or intersection
-	 */
-	Container,
-
-	/**
-	 * Type reference created during type checking
-	 * @description Usually Array<...>, ReadOnly<...> etc.
-	 */
-	TransientTypeReference,
-
-	/**
-	 * Some specific object
-	 * @description Eg. "{ foo: string, bar: boolean }"
-	 */
-	Object,
-
-	/**
-	 * Some subtype of string, number, boolean
-	 * @example <caption>type Foo = "hello world" | "hello"</caption>
-	 * String "hello world" is literal type and it is subtype of string.
-	 *
-	 * <caption>type TheOnlyTrue = true;</caption>
-	 * Same as true is literal type and it is subtype of boolean.
-	 */
-	LiteralType,
-
-	/**
-	 * Fixed lenght arrays literals
-	 * @example <caption>type Coords = [x: number, y: number, z: number];</caption>
-	 */
-	Tuple,
-
-	/**
-	 * Generic parameter type
-	 * @description Represent generic type parameter of generic types. Eg. it is TType of class Animal<TType> {}.
-	 */
-	TypeParameter,
-
-	/**
-	 * Conditional type
-	 */
-	ConditionalType = 9,
-
-	/**
-	 * Indexed access type
-	 * @description Eg. get<K extends keyof TypeKind>(key: K): ==>> TypeKind[K] <<==
-	 */
-	IndexedAccess = 10,
-
-	/**
-	 * Typescript "module"
-	 * @description Value module or namespace module
-	 */
-	Module = 11,
-
-	/**
-	 * Specific method used as type
-	 */
-	Method = 12,
-}
-
-export enum Accessor
-{
-	None,
-	Getter,
-	Setter
-}
-
-export enum AccessModifier
-{
-	Private,
-	Protected,
-	Public
-}
-
-/**
- * @internal
- */
-export interface ConditionalTypeDescription
-{
-	/**
-	 * Extends type
-	 */
-	e: Type;
-
-	/**
-	 * True type
-	 */
-	tt: Type;
-
-	/**
-	 * False type
-	 */
-	ft: Type;
-}
-
-export interface ConditionalType
-{
-	/**
-	 * Extends type
-	 */
-	extends: Type;
-
-	/**
-	 * True type
-	 */
-	trueType: Type;
-
-	/**
-	 * False type
-	 */
-	falseType: Type;
-}
-
-/**
- * @internal
- */
-export interface IndexedAccessTypeDescription
-{
-	/**
-	 * Object type
-	 */
-	ot: Type;
-
-	/**
-	 * Index type
-	 */
-	it: Type;
-}
-
-export interface IndexedAccessType
-{
-	/**
-	 * Object type
-	 */
-	objectType: Type;
-
-	/**
-	 * Index type
-	 */
-	indexType: Type;
-}
-
-/**
- * @internal
- */
-export interface ParameterDescription
-{
-	/**
-	 * Name of the parameter
-	 */
-	n: string;
-
-	/**
-	 * Type of the parameter
-	 */
-	t: Type;
-
-	/**
-	 * Optional parameter
-	 */
-	o: boolean;
-}
-
-/**
- * @internal
- */
-export interface PropertyDescription
-{
-	/**
-	 * Name of the property
-	 */
-	n: string;
-
-	/**
-	 * Property type
-	 */
-	t: Type;
-
-	/**
-	 * Optional property
-	 */
-	o: boolean;
-
-	/**
-	 * Decorators
-	 */
-	d?: Array<DecoratorDescription>;
-
-	/**
-	 * Access modifier
-	 */
-	am?: AccessModifier;
-
-	/**
-	 * Accessor
-	 */
-	acs?: Accessor;
-
-	/**
-	 * Readonly
-	 */
-	ro?: boolean;
-}
-
-/**
- * Property description
- */
-export interface Property
-{
-	/**
-	 * Property name
-	 */
-	name: string;
-
-	/**
-	 * Property type
-	 */
-	type: Type;
-
-	/**
-	 * Optional property
-	 */
-	optional: boolean;
-
-	/**
-	 * Property decorators
-	 */
-	decorators: ReadonlyArray<Decorator>;
-
-	/**
-	 * Access modifier
-	 */
-	accessModifier: AccessModifier;
-
-	/**
-	 * Accessor
-	 */
-	accessor: Accessor;
-
-	/**
-	 * Readonly
-	 */
-	readonly: boolean;
-}
-
-/**
- * @internal
- */
-export interface DecoratorDescription
-{
-	n: string;
-	fn: string;
-}
-
-/**
- * Decoration description
- */
-export interface Decorator
-{
-	/**
-	 * Decorator name
-	 */
-	name: string;
-
-	/**
-	 * Decorator full name
-	 */
-	fullName?: string;
-}
-
-/**
- * Method parameter description
- */
-export interface MethodParameter
-{
-	/**
-	 * Parameter name
-	 */
-	name: string;
-
-	/**
-	 * Parameter type
-	 */
-	type: Type;
-
-	/**
-	 * Parameter is optional
-	 */
-	optional: boolean;
-}
-
-/**
- * @internal
- */
-export interface MethodDescription
-{
-	/**
-	 * Method name
-	 */
-	n: string;
-
-	/**
-	 * Parameters
-	 */
-	params: Array<ParameterDescription>;
-
-	/**
-	 * Return type
-	 */
-	rt: Type;
-
-	/**
-	 * Decorators
-	 */
-	d?: Array<DecoratorDescription>;
-
-	/**
-	 * Generic type parameters
-	 */
-	tp?: Array<Type>;
-
-	/**
-	 * Optional method
-	 */
-	o: boolean;
-
-	/**
-	 * Access modifier
-	 */
-	am: AccessModifier;
-}
-
-const Mapper = {
-	/**
-	 * @internal
-	 * @param d
-	 */
-	mapDecorators(d: DecoratorDescription): Decorator
-	{
-		return ({ name: d.n, fullName: d.fn });
-	},
-
-	/**
-	 * @internal
-	 * @param p
-	 */
-	mapProperties(p: PropertyDescription): Property
-	{
-		return ({
-			name: p.n,
-			type: resolveLazyType(p.t),
-			decorators: p.d?.map(Mapper.mapDecorators) || [],
-			optional: p.o,
-			accessModifier: p.am ?? AccessModifier.Public,
-			accessor: p.acs ?? Accessor.None,
-			readonly: p.ro ?? false
-		});
-	},
-
-	/**
-	 * @internal
-	 * @param c
-	 */
-	mapConstructors(c: ConstructorDescription): Constructor
-	{
-		return Reflect.construct(Constructor, [c], ConstructorActivator);
-	},
-
-	/**
-	 * @internal
-	 * @param p
-	 * @return {{name: string, optional: boolean, type: Type}}
-	 */
-	mapMethodParameters(p: ParameterDescription): MethodParameter
-	{
-		return ({
-			name: p.n,
-			type: resolveLazyType(p.t),
-			optional: p.o
-		});
-	}
-};
-
-export class MethodBase
-{
-	private readonly _parameters: Array<MethodParameter>;
-
-	/**
-	 * Internal constructor
-	 * @internal
-	 */
-	protected constructor(params: Array<ParameterDescription>)
-	{
-		this._parameters = params.map(Mapper.mapMethodParameters);
-	}
-
-	/**
-	 * Parameters of this method
-	 */
-	getParameters(): ReadonlyArray<MethodParameter>
-	{
-		return this._parameters.slice();
-	}
-}
-
-/**
- * Method details
- */
-export class Method extends MethodBase
-{
-	private readonly _name: string;
-	private readonly _returnType: Type;
-	private readonly _optional: boolean;
-	private readonly _typeParameters: Array<Type>;
-	private readonly _decorators: Array<Decorator>;
-	private readonly _accessModifier: AccessModifier;
-
-	/**
-	 * Name of this method
-	 */
-	get name(): string
-	{
-		return this._name;
-	}
-
-	/**
-	 * Return type of this method
-	 */
-	get returnType(): Type
-	{
-		return this._returnType;
-	}
-
-	/**
-	 * Method is optional
-	 */
-	get optional(): boolean
-	{
-		return this._optional;
-	}
-
-	/**
-	 * Access modifier
-	 */
-	get accessModifier(): AccessModifier
-	{
-		return this._accessModifier;
-	}
-
-	/**
-	 * Internal method constructor
-	 * @internal
-	 */
-	constructor(description: MethodDescription)
-	{
-		super(description.params);
-
-		if (new.target != MethodActivator)
-		{
-			throw new Error("You cannot create instance of Method manually!");
-		}
-
-		this._name = description.n;
-		this._typeParameters = description.tp?.map(t => resolveLazyType(t)) || [];
-		this._returnType = resolveLazyType(description.rt);
-		this._optional = description.o;
-		this._accessModifier = description.am;
-		this._decorators = description.d?.map(Mapper.mapDecorators) || [];
-	}
-
-	/**
-	 * Returns list of generic type parameter.
-	 * @return {Array<Type>}
-	 */
-	getTypeParameters(): ReadonlyArray<Type>
-	{
-		return this._typeParameters.slice();
-	}
-
-	/**
-	 * Returns array of decorators
-	 */
-	getDecorators(): ReadonlyArray<Decorator>
-	{
-		return this._decorators.slice();
-	}
-}
-
-class MethodActivator extends Method
-{
-}
-
-/**
- * @internal
- */
-export interface ConstructorDescription
-{
-	params: Array<ParameterDescription>;
-}
-
-/**
- * Constructor details
- */
-export class Constructor extends MethodBase
-{
-	/**
-	 * Internal constructor
-	 * @internal
-	 */
-	constructor(description: ConstructorDescription)
-	{
-		super(description.params);
-
-		if (new.target != ConstructorActivator)
-		{
-			throw new Error("You cannot create instance of Constructor manually!");
-		}
-	}
-}
-
-class ConstructorActivator extends Constructor
-{
-}
-
-/**
- * This data is not set when the config mode is set to "universal"
- * @internal
- */
-export interface ConstructorImportDescription
-{
-	/**
-	 * This is the name of the actual declaration
-	 * In the example above, this would be "SomeClass"
-	 */
-	n: string | undefined;
-	/**
-	 * The exported name of this constructor from its source file.
-	 * For example;
-	 * "export class SomeClass {}" would be "SomeClass"
-	 * "export default class SomeClass {}" would be "default"
-	 */
-	en: string | undefined;
-	/**
-	 * The absolute path of the source file for this constructor
-	 */
-	srcPath: string | undefined;
-	/**
-	 * The absolute path for the javascript file of this constructor
-	 */
-	outPath: string | undefined;
-}
-
-/**
- * Method details
- */
-export class ConstructorImport
-{
-	private readonly _name: string | undefined;
-	private readonly _exportName: string | undefined;
-	private readonly _sourcePath: string | undefined;
-	private readonly _outPath: string | undefined;
-
-	/**
-	 * @internal
-	 */
-	constructor(description: ConstructorImportDescription | undefined)
-	{
-		if (new.target != ConstructorImportActivator)
-		{
-			throw new Error("You cannot create instance of Method manually!");
-		}
-
-		if (!description)
-		{
-			return;
-		}
-		this._name = description.n;
-		this._exportName = description.en;
-		this._sourcePath = description.srcPath;
-		this._outPath = description.outPath;
-	}
-
-	/**
-	 * This is the name of the actual declaration
-	 * In the example above, this would be "SomeClass"
-	 */
-	get name(): string | undefined
-	{
-		return this._name;
-	}
-
-	/**
-	 * The exported name of this constructor from its source file.
-	 * For example;
-	 * "export class SomeClass {}" would be "SomeClass"
-	 * "export default class SomeClass {}" would be "default"
-	 */
-	get exportName(): string | undefined
-	{
-		return this._exportName;
-	}
-
-	/**
-	 * The absolute path of the source file for this constructor
-	 */
-	get sourcePath(): string | undefined
-	{
-		return this._sourcePath;
-	}
-
-	/**
-	 * The absolute path for the javascript file of this constructor
-	 */
-	get outPath(): string | undefined
-	{
-		return this._outPath;
-	}
-}
-
-class ConstructorImportActivator extends ConstructorImport
-{
-}
-
-/**
- * @internal
- */
-export interface TypeProperties
-{
-	/**
-	 * Type name
-	 */
-	n?: string;
-
-	/**
-	 * Type fullname
-	 */
-	fn?: string;
-
-	/**
-	 * TypeKind
-	 */
-	k: TypeKind;
-
-	/**
-	 * Constructors
-	 */
-	ctors?: Array<ConstructorDescription>;
-
-	/**
-	 * Properties
-	 */
-	props?: Array<PropertyDescription>;
-
-	/**
-	 * Methods
-	 */
-	meths?: Array<MethodDescription>;
-
-	/**
-	 * Decorators
-	 */
-	decs?: Array<DecoratorDescription>;
-
-	/**
-	 * Generic type parameters
-	 */
-	tp?: Array<Type>;
-
-	/**
-	 * Is union type
-	 */
-	union?: boolean;
-
-	/**
-	 * Is intersection type
-	 */
-	inter?: boolean;
-
-	/**
-	 * Unified or intersecting types
-	 */
-	types?: Array<Type>;
-
-	/**
-	 * Some path/export information about the module it-self. Will help dynamically import modules.
-	 * This data is not set when the config mode is set to "universal"
-	 */
-	ctorDesc?: ConstructorImportDescription;
-
-	/**
-	 * Ctor getter
-	 */
-	ctor?: () => Function;
-
-	/**
-	 * Extended base type
-	 */
-	bt?: Type;
-
-	/**
-	 * Implemented interface
-	 */
-	iface?: Type;
-
-	/**
-	 * Literal value
-	 */
-	v?: any;
-
-	/**
-	 * Type arguments
-	 */
-	args?: Array<Type>;
-
-	/**
-	 * Default type
-	 */
-	def?: Type;
-
-	/**
-	 * Constraining type
-	 */
-	con?: Type;
-
-	/**
-	 * Conditional type description
-	 */
-	ct?: ConditionalTypeDescription;
-
-	/**
-	 * Indexed access type description
-	 */
-	iat?: IndexedAccessTypeDescription;
-}
 
 /**
  * Object representing TypeScript type in memory
@@ -835,7 +97,7 @@ export class Type
 		this._kind = description.k;
 		this._constructors = description.ctors?.map(Mapper.mapConstructors) || [];
 		this._properties = description.props?.map(Mapper.mapProperties) || [];
-		this._methods = description.meths?.map(m => Reflect.construct(Method, [m], MethodActivator)) || [];
+		this._methods = description.meths?.map(Mapper.mapMethods) || [];
 		this._decorators = description.decs?.map(Mapper.mapDecorators) || [];
 		this._typeParameters = description.tp?.map(t => resolveLazyType(t)) || [];
 		this._ctor = description.ctor;
@@ -1352,6 +614,7 @@ export class TypeActivator extends Type
 export function getType<T>(): Type | undefined
 // TODO: Uncomment and use this line. Waiting for TypeScript issue to be resolved. https://github.com/microsoft/TypeScript/issues/46155
 // export function getType<T = void>(..._: (T extends void ? ["You must provide a type parameter"] : [])): Type | undefined
+
 export function getType<T>(): Type | undefined
 {
 	return undefined;
@@ -1370,18 +633,6 @@ getType.lazy = function (typeId: number) {
 	};
 };
 
-/**
- * @internal
- */
-function resolveLazyType(type?: Type | Function)
-{
-	if (typeof type == "function" && type.name == "lazyType")
-	{
-		return type();
-	}
-
-	return type;
-}
 
 /**
  * Class decorator which marks classes to be processed and included in metadata lib file.
@@ -1393,33 +644,4 @@ export function reflect<TType>()
 	return function <T>(Constructor: { new(...args: any[]): T }) {
 	};
 }
-
-/**
- * To identify functions by package
- */
-export const TYPE_ID_PROPERTY_NAME = "__tst_reflect__";
-
-/**
- * Name of JSDoc comment marking method as it use type of generic parameter.
- */
-export const REFLECT_GENERIC_DECORATOR = "reflectGeneric";
-
-/**
- * Name of JSDoc comment marking function as decorator.
- * @type {string}
- */
-export const REFLECT_DECORATOR_DECORATOR = "reflectDecorator";
-
-/**
- * Name of the getType() function
- */
-export const GET_TYPE_FNC_NAME = "getType";
-
-/**
- * Name of the getType.lazy() function
- * @type {string}
- */
-export const GET_TYPE_LAZY_FNC_NAME = "lazy";
-
-export const REFLECT_STORE_SYMBOL = Symbol('tst_reflect_store');
