+++ transformer/src/getTypeCall.ts
@@ -1,5 +1,4 @@
 import * as ts                    from "typescript";
-import { GET_TYPE_LAZY_FNC_NAME } from "tst-reflect";
 import {
 	GetTypeCall,
 	TypePropertiesSource
@@ -50,15 +49,6 @@ export function getTypeCall(type: ts.Type, symbol: ts.Symbol | undefined, contex
 			if (creatingTypes.includes(id))
 			{
 				return context.metaWriter.factory.getTypeFromStoreLazily(id);
-
-				/**
-				 * Original in file lazy expression
-				 */
-				// return ts.factory.createCallExpression(
-				// 	ts.factory.createPropertyAccessExpression(getTypeIdentifier, GET_TYPE_LAZY_FNC_NAME),
-				// 	[],
-				// 	[ts.factory.createNumericLiteral(id)]
-				// );
 			}
 
 			creatingTypes.push(id);
@@ -81,25 +71,13 @@ export function getTypeCall(type: ts.Type, symbol: ts.Symbol | undefined, contex
 		/**
 		 * Just call `getType()` with typeId; Type is going to be loaded from storage
 		 */
-
 		return context.metaWriter.factory.getTypeFromStore(id);
-
-		/**
-		 * Original in file getType(id) call
-		 */
-		// return ts.factory.createCallExpression(getTypeIdentifier, [], [ts.factory.createNumericLiteral(id)]);
 	}
 
 	/**
 	 * Type is not registered (no id or no sourceFileContext) so direct type construction returned
 	 */
-
 	return context.metaWriter.factory.createDescriptionWithoutAddingToStore(props);
-
-	/**
-	 * Original in file getType({...}) call
-	 */
-	// return ts.factory.createCallExpression(getTypeIdentifier, [], [typePropertiesObjectLiteral]);
 }
 
 /**
