+++ transformer/src/meta-writer/impl/index.ts
@@ -1,3 +1,2 @@
-export * from './MetaStoreImpl';
 export * from './MetaLibTransformerImpl';
 export * from './MetaWriterNodeGeneratorImpl';
